requirements.txt
tensorflow
python-telegram-bot
Pillow
numpy